feat: refactor A2UI catalog structure to versioned paths and enhance catalog assembly
