+++ packages/core/src/a2ui/v0_9/block-catalog-json.ts
