+++ packages/core/src/a2ui/v0_9/catalog-assembler.ts
