+++ packages/core/src/index.ts
@@ -1,7 +1,10 @@
-export * from "./a2ui/block-catalog";
+export * from "./a2ui/v0_9/block-catalog";
 export {
   assembleCatalog,
   type AssembledCatalog,
   type CatalogEnvelope,
-} from "./a2ui/catalog-assembler";
-export { BLOCK_CATALOG_ID, blockCatalogJson } from "./a2ui/block-catalog-json";
+} from "./a2ui/v0_9/catalog-assembler";
+export {
+  BLOCK_CATALOG_ID,
+  blockCatalogJson,
+} from "./a2ui/v0_9/block-catalog-json";
